refactor(entity-ui): extract tool-list management into useAgentTools
Move the agent-template Tools add/remove logic (inline function tools,
builtin/gateway tools, async workflow-reference tools, and the derived
selected-name set + referenceable-workflow pool) out of AgentTemplateControl
into a useAgentTools hook. The orchestrator now just wires the handlers into
the picker and renders.

AgentTemplateControl drops to 580 lines (from ~2.1k before the split);
behavior is unchanged.
